apps/worker/requirements.txt
livekit-agents>=1.0.0
livekit-plugins-openai>=1.0.0
livekit-plugins-google>=1.0.0
cryptography>=43.0
psycopg[binary]>=3.1
pydantic>=2.0
python-dotenv>=1.0
pyyaml>=6.0
